old-tests/submission/W3C/domhtml/HTMLParagraphElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - PARAGRAPH</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-p-align" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p align="center">
TEXT
</p>



<script>
var domtest = async_test("The align IDL attribute of the p element must reflect the content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var valign;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("p");
      assert_equals(nodeList.length,2);
testNode = nodeList.item(0);
      valign = testNode.align;

      assert_equals(valign,"center");
       
    domtest.done();
}));

</script>
<p>The align IDL attribute of the p element must reflect the content attribute of the same name.</p>
<div id="log"></div>
</body></html>
